Renamed "relative.rs" to "source_rectangled.rs"
