+++ src/source_rectangled.rs
